% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penet_large.R
\docType{data}
\name{penet_large}
\alias{penet_large}
\title{A penetrance matrix relating the phenotypes in \code{dat_large} to three genotypes}
\format{
A matrix with 10,002 rows (corresponding to persons) and 3
columns (corresponding to genotypes).
}
\source{
Simulated
}
\usage{
penet_large
}
\description{
A matrix relating the phenotypes of \code{dat_large} to the three unphased
genotypes of a single biallelic, autosomal genetic locus.
The element \code{penet_large[i,j]} is the conditional probability of
the phenotypes (i.e. \code{sex}, \code{aff} and \code{age}) of the person
in row \code{i} of \code{dat_large}, given that
his or her genotype is \code{j} (here labelling the genotypes as \verb{1, 2, 3},
where genotype \code{2} is the heterozygous genotype).
}
\keyword{datasets}
