% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penet_small.R
\docType{data}
\name{penet_small}
\alias{penet_small}
\title{A penetrance matrix relating the phenotypes of \code{dat_small} to three genotypes}
\format{
A matrix with 1018 rows (corresponding to persons) and 3
columns (corresponding to genotypes).
}
\source{
Simulated
}
\usage{
penet_small
}
\description{
A matrix relating the phenotypes of \code{dat_small} to the three unphased
genotypes of a single biallelic, autosomal genetic locus.
The element \code{penet_small[i,j]} is the conditional probability of the
the phenotypes (i.e. \code{sex}, \code{aff} and \code{age}) of the person
in row \code{i} of \code{dat_small}, given that
his or her genotype is \code{j} (here labelling the genotypes as \verb{1, 2, 3},
where genotype \code{2} is the heterozygous genotype)
}
\keyword{datasets}
