% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_small.R
\docType{data}
\name{dat_small}
\alias{dat_small}
\title{Simulated data on 10 families with approximately 100 members each}
\format{
A data frame with 1018 rows (corresponding to persons) and the
following 7 variables:
\describe{
\item{\code{family}}{an identifier for each person's family}
\item{\code{indiv}}{an individual identifier for each person}
\item{\code{mother}}{the individual identifier of the person's mother}
\item{\code{father}}{the individual identifier of the person's father}
\item{\code{sex}}{the person's sex (1 = male, 2 = female)}
\item{\code{aff}}{the person's disease status (1 = case, 0 = control)}
\item{\code{age}}{the person's last known age in years (for controls) or
age at diagnosis in years (for cases)}
\item{\code{geno}}{the person's genotype, or blank (\code{""}) if not known}
}
}
\source{
Simulated
}
\usage{
dat_small
}
\description{
A dataset giving the relationship structure of 10 families
and phenotypic data on the family members
}
\keyword{datasets}
